+++ libs/openFrameworks/graphics/ofPolyline.inl
@@ -406,7 +406,7 @@ void ofPolyline_<T>::arc(const T & center, float radiusX, float radiusY, float a
         if(clockwise) {
             currentLUTIndex = currentLUTIndex % nCirclePoints;
         } else {
-            if(currentLUTIndex < 0) currentLUTIndex = nCirclePoints + currentLUTIndex;
+            if(currentLUTIndex < 0) currentLUTIndex = static_cast<int>(nCirclePoints + currentLUTIndex);
         }
         
         // add the point to the poly line
@@ -756,7 +756,7 @@ template<class T>
 void ofPolyline_<T>::simplify(float tol){
     if(points.size() < 2) return;
     
-	int n = size();
+	int n = static_cast<int>(size());
 	
 	if(n == 0) {
 		return;
@@ -893,7 +893,7 @@ float ofPolyline_<T>::getIndexAtLength(float length) const {
     float totalLength = getPerimeter();
     length = ofClamp(length, 0, totalLength);
     
-    int lastPointIndex = isClosed() ? points.size() : points.size()-1;
+    int lastPointIndex = static_cast<int>(isClosed() ? points.size() : points.size()-1);
     
     int i1 = ofClamp(floor(length / totalLength * lastPointIndex), 0, lengths.size()-2);   // start approximation here
     int leftLimit = 0;
@@ -1120,8 +1120,8 @@ template<class T>
 int ofPolyline_<T>::getWrappedIndex(int index) const {
     if(points.empty()) return 0;
     
-    if(index < 0) return isClosed() ? (index + points.size()) % points.size() : 0;
-    if(index > int(points.size())-1) return isClosed() ? index % points.size() : points.size() - 1;
+    if(index < 0) return static_cast<int>(isClosed() ? (index + points.size()) % points.size() : 0);
+    if(index > int(points.size())-1) return static_cast<int>(isClosed() ? index % points.size() : points.size() - 1);
     return index;
 }
